pyproject.toml
[project]
name = "blogagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.23",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.12",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.3",
    "python-dotenv>=1.2.2",
]
